💡 chore(users): response 파일 위치 변경 및 미사용 import 제거
